Rename insight-classes to insight-objects
