+++ packages/server/src/insights/insight-objects.ts
