+++ packages/server/test/insights.integration.ts
@@ -8,7 +8,7 @@ import {
 import { setupIntegrationTest } from './util/testUtils';
 import { InsightsModule } from '../src/insights/insights.module';
 import { Role } from '@koh/common';
-import { TotalQuestionsAsked } from 'insights/insight-classes';
+import { TotalQuestionsAsked } from 'insights/insight-objects';
 
 describe('Insights Integration', () => {
   const supertest = setupIntegrationTest(InsightsModule);
